Move parsers into src/ftl
