+++ src/ftl/ast/parser.js
@@ -1,7 +1,7 @@
 /*eslint no-magic-numbers: [0]*/
 
 import AST from './ast';
-import { L10nError } from '../../../errors';
+import { L10nError } from '../../lib/errors';
 
 const MAX_PLACEABLES = 100;
