+++ src/ftl/entries/transformer.js
