+++ src/runtime/node/index.js
@@ -1,12 +1,9 @@
 import 'string.prototype.startswith';
 import 'string.prototype.endswith';
 
-export { default as FTLASTParser } from
-  '../../lib/format/ftl/ast/parser';
-export { default as FTLEntriesParser } from
-  '../../lib/format/ftl/entries/parser';
-export { createEntriesFromAST } from
-  '../../lib/format/ftl/entries/transformer';
+export { default as FTLASTParser } from '../../ftl/ast/parser';
+export { default as FTLEntriesParser } from '../../ftl/entries/parser';
+export { createEntriesFromAST } from '../../ftl/entries/transformer';
 export { Bundle } from '../../intl/bundle';
 
 import { Context, SimpleContext } from '../../lib/context';
