+++ src/runtime/tooling/index.js
@@ -1,9 +1,9 @@
 import { fetchResource } from '../web/io';
 import { View } from '../../bindings/html/view';
 
-import FTLASTParser from '../../lib/format/ftl/ast/parser';
-import FTLEntriesParser from '../../lib/format/ftl/entries/parser';
-import {createEntriesFromAST} from '../../lib/format/ftl/entries/transformer';
+import FTLASTParser from '../../ftl/ast/parser';
+import FTLEntriesParser from '../../ftl/entries/parser';
+import {createEntriesFromAST} from '../../ftl/entries/transformer';
 
 import { Context, SimpleContext } from '../../lib/context';
 import { L10nError } from '../../lib/errors';
