+++ tools/parse.js
@@ -11,9 +11,9 @@ require('babel-register')({
   plugins: ['transform-es2015-modules-commonjs']
 });
 
-const FTLASTParser = require('../src/lib/format/ftl/ast/parser').default;
-const FTLRuntimeParser = require('../src/lib/format/ftl/entries/parser').default;
-const { createEntriesFromAST } = require('../src/lib/format/ftl/entries/transformer');
+const FTLASTParser = require('../src/ftl/ast/parser').default;
+const FTLRuntimeParser = require('../src/ftl/entries/parser').default;
+const { createEntriesFromAST } = require('../src/ftl/entries/transformer');
 
 program
   .version('0.0.1')
